Address compiler warnings/errors

See merge request jvet/VVCSoftware_VTM!2361

+++ source/Lib/CommonLib/Quant.cpp
@@ -1104,7 +1104,8 @@ void Quant::transformSkipQuantOneSample(TransformUnit &tu, const ComponentID &co
   */
   const int iQBits = QUANT_SHIFT + cQP.per(useTransformSkip) + (useTransformSkip ? 0 : iTransformShift);
   // QBits will be OK for any internal bit depth as the reduction in transform shift is balanced by an increase in Qp_per due to QpBDOffset
-  const int iAdd = int64_t(bUseHalfRoundingPoint ? 256 : (tu.cs->slice->isIRAP() ? 171 : 85)) << int64_t(iQBits - 9);
+  const int64_t iAdd = int64_t(bUseHalfRoundingPoint ? 256 : (tu.cs->slice->isIRAP() ? 171 : 85))
+                       << int64_t(iQBits - 9);
   TCoeff transformedCoefficient;
 
   // transform-skip
